requirements.txt
# ============================================================
# LANGCHAIN SECURITY RAG — Python Dependencies
# Install with: pip install -r requirements.txt
# ============================================================

# LangChain Core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0

# Vector Database
chromadb>=0.5.0

# PDF Processing
pypdf>=3.0.0

# Embeddings
sentence-transformers>=2.2.0

# Web UI
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Image processing (ESL preview + dithering)
Pillow>=10.0.0
numpy>=1.24.0
